+++ src/data/equipment.js
@@ -1,20 +1,27 @@
+const SWAPS = [
+  /^(Greater) (.*)/,
+  /^(Lesser) (.*)/,
+  /^(Major) (.*)/,
+  /^(Moderate) (.*)/,
+  /^(Standard) (.*)/,
+];
+
 // this equipment is named differently in foundry vs pathbuilder
-export const EQUIPMENT_RENAME_MAP = [
+export const EQUIPMENT_RENAME_STATIC_MAP = [
   { pbName: "Chain", foundryName: "Chain (10 feet)" },
   { pbName: "Oil", foundryName: "Oil (1 pint)" },
   { pbName: "Bracelets of Dashing", foundryName: "Bracelet of Dashing" },
   { pbName: "Fingerprinting Kit", foundryName: "Fingerprint Kit" },
-  { pbName: "Greater Unmemorable Mantle", foundryName: "Unmemorable Mantle (Greater)" },
-  { pbName: "Major Unmemorable Mantle", foundryName: "Unmemorable Mantle (Major)" },
   { pbName: "Ladder", foundryName: "Ladder (10-foot)" },
   { pbName: "Mezmerizing Opal", foundryName: "Mesmerizing Opal" },
   { pbName: "Explorer's Clothing", foundryName: "Clothing (Explorer's)" },
   { pbName: "Flaming Star (Greater)", foundryName: "Greater Flaming Star" },
   { pbName: "Potion of Lesser Darkvision", foundryName: "Darkvision Elixir (Lesser)" },
+  { pbName: "Potion of Greater Darkvision", foundryName: "Darkvision Elixir (Greater)" },
+  { pbName: "Potion of Moderate Darkvision", foundryName: "Darkvision Elixir (Moderate)" },
   { pbName: "Bottled Sunlight", foundryName: "Formulated Sunlight" },
   { pbName: "Magazine (Repeating Hand Crossbow)", foundryName: "Magazine with 5 Bolts" },
   { pbName: "Astrolabe (Standard)", foundryName: "Standard Astrolabe" },
-  { pbName: "Greater Cloak of Repute", foundryName: "Cloak of Repute (Greater)" },
   { pbName: "Skinitch Salve", foundryName: "Skinstitch Salve" },
   { pbName: "Flawless Scale", foundryName: "Abadar's Flawless Scale" },
   { pbName: "Construct Key", foundryName: "Cordelia's Construct Key" },
@@ -26,6 +33,25 @@ export const EQUIPMENT_RENAME_MAP = [
   { pbName: "Heartstone", foundryName: "Skarja's Heartstone" },
   { pbName: "Bullets (10 rounds)", foundryName: "Sling Bullets" },
 ];
+
+function generateDynamicNames(pbName) {
+  const result = [];
+  // if we have a hardcoded map, don't return here
+  for (const reg of SWAPS) {
+    const match = pbName.match(reg);
+    if (match) {
+      result.push({ pbName, foundryName: `${match[2]} (${match[1]})`, details: match[2] });
+    }
+  }
+  return result;
+}
+
+
+export function EQUIPMENT_RENAME_MAP(pbName = null) {
+  const postfixNames = pbName ? generateDynamicNames(pbName) : [];
+  return postfixNames.concat(EQUIPMENT_RENAME_STATIC_MAP);
+}
+
 
 // this is equipment is special and shouldn't have the transformations applied to it
 export const RESTRICTED_EQUIPMENT = [